requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
black==19.10b0
click==7.1.2
Django==3.0.3
django-countries==6.1.2
django-phonenumber-field==4.0.0
extras==1.0.0
linecache2==1.0.0
parse==1.15.0
parse-type==0.5.2
pathspec==0.8.0
pbr==5.4.5
phonenumbers==8.12.2
Pillow==7.1.2
pylint==2.5.0
pre-commit==2.3.0
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
testtools==2.4.0
toml==0.10.0
traceback2==1.4.0
typed-ast==1.4.1
unittest2==1.1.0
urllib3==1.25.9
gunicorn==20.0.4
psycopg2-binary==2.8.5